Add a class for stats.
This class only exposes the stats of a specific translation
and not of a whole resource. The resource stats can be
fetched by querying the collection url instead of the item
url that the default get() method uses. Adding an all() method
should solve that, but it would break the design of the
library, as an all() method should propably return a list of
Stats objects and not a huge object.