Remove @property decorators from BaseModel
__getattr__ and @property do not mix well. If a property raises
an AttributeError, Python (`__getattribute__()`) falls back to using
`__getattr__()`. The problem is that the traceback is lost, which makes
it extremely hard to debug, since the actual exception raised
isn't shown anywhere.

This could happen during testing, where some functionality was mocked,
when trying to save a new object, e.g. a `Resource`. If not all
necessary URL fields were defined in the `Resource` constructor
(e.g. if `slug` wasn't defined) then the `url_parameters` property
would raise an exception, but the actual exception that would be shown
was "Resource has no readable attribute url_properties`, raised by
`__getattr__`. This makes no sense, since `url_properties` is
an actual member of `Resource`, thus debugging was hard.

Since most of the functionality of `BaseModel` is based on
`__getattr__`, the only way to fix this easily was to remove all
properties.